requirements.txt
# --- Core Computer Vision & Processing ---
# Using contrib version for ArUco markers and advanced algorithms
opencv-contrib-python>=4.8.0
numpy>=1.24.0

# --- Testing & BDD Framework ---
pytest>=7.4.0
pytest-bdd>=6.1.1
# Required for modern dependency checking in tests
setuptools
packaging

# --- Utilities ---
# For loading camera calibration files (US-14)
pyyaml>=6.0.1
# Helpful for path manipulations
pathlib>=1.0.1